+++ cedar-policy/src/proto/policy.rs
@@ -354,14 +354,9 @@ impl From<&ast::ActionConstraint> for models::ActionConstraint {
                 )),
             },
             #[cfg(feature = "tolerant-ast")]
-            ast::ActionConstraint::ErrorConstraint =>
-            // Treat an error constraint as an Any constraint for Protobufs since Protobufs schema model has no Error
-            {
-                Self {
-                    data: Some(models::action_constraint::Data::Any(
-                        models::action_constraint::Any::Unit.into(),
-                    )),
-                }
+            #[expect(clippy::unimplemented, reason = "experimental feature")]
+            ast::ActionConstraint::ErrorConstraint => {
+                unimplemented!("tolerant-ast cannot be used with the protobuf feature")
             }
         }
     }